recipe-manager/pyproject.toml
[project]
name = "recipe-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.1",
    "httpx>=0.28.1",
    "libsql-client>=0.3.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.0",
]
